api/requirements.txt
# LuckyRed API Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.13.0
loguru>=0.7.0
python-telegram-bot>=20.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
